+++ plume/src/main/kotlin/io/github/plume/oss/util/ExtractorConst.kt
@@ -1,6 +1,8 @@
 package io.github.plume.oss.util
 
 object ExtractorConst {
+    const val LANGUAGE_FRONTEND = "Plume"
+    const val LANGUAGE_FRONTEND_VERSION = "0.1"
     val PRIMITIVES = listOf("boolean", "byte", "char", "double", "float", "int", "long", "null", "short", "void")
     val BIN_OPS: Map<String, String> = mapOf(
         "+" to "ADD",