+++ plume/src/test/kotlin/io/github/plume/oss/drivers/OverflowDbDriverIntTest.kt
@@ -225,6 +225,16 @@ class OverflowDbDriverIntTest {
             driver.getWholeGraph()
                 .use { g -> assertTrue(g.nodes(fileVertex.id()).asSequence().any { it.property(NAME) == STRING_2 }) }
         }
+
+        @Test
+        fun testGetMetaData() {
+            driver.addVertex(metaDataVertex)
+            val metaData = driver.getMetaData()
+            assertNotNull(metaData)
+            assertEquals(metaDataVertex.id(), metaData!!.id())
+            driver.deleteVertex(metaData.id(), metaData.build().label())
+            assertNull(driver.getMetaData())
+        }
     }
 
     @Nested
@@ -536,10 +546,12 @@ class OverflowDbDriverIntTest {
 
         @Test
         fun testGetProgramStructure() {
+            val unknown = io.shiftleft.semanticcpg.language.types.structure.File.UNKNOWN()
+            driver.addVertex(NewFileBuilder().name(unknown).order(0).hash(Option.apply(unknown)))
             g = driver.getProgramStructure()
             val ns = g.nodes().asSequence().toList()
             val es = g.edges().asSequence().toList()
-            assertEquals(4, ns.size)
+            assertEquals(5, ns.size)
             assertEquals(2, es.size)
 
             val file = g.V(fileVertex.id()).next()