+++ plume/src/test/kotlin/io/github/plume/oss/extractor/interprocedural/TypeInterproceduralTest.kt
@@ -39,8 +39,7 @@ class TypeInterproceduralTest {
         val resourceDir = "${PATH.absolutePath}${File.separator}Type$currentTestNumber.java"
         // Load test resource and project + export graph
         val f = File(resourceDir)
-        extractor.load(f)
-        extractor.project()
+        extractor.load(f).project().postProject()
         g = driver.getWholeGraph()
     }