+++ plume/src/test/kotlin/io/github/plume/oss/extractor/intraprocedural/BasicIntraproceduralTest.kt
@@ -43,8 +43,7 @@ class BasicIntraproceduralTest {
         val resourceDir = "${PATH.absolutePath}${File.separator}Basic$currentTestNumber.java"
         // Load test resource and project + export graph
         val f = File(resourceDir)
-        extractor.load(f)
-        extractor.project()
+        extractor.load(f).project().postProject()
         g = driver.getWholeGraph()
     }