+++ plume/src/test/kotlin/io/github/plume/oss/extractor/intraprocedural/SwitchIntraproceduralTest.kt
@@ -45,8 +45,7 @@ class SwitchIntraproceduralTest {
         val testSourceFile = "$testResourcePath.java"
         // Load test resource and project + export graph
         val f = File(testSourceFile)
-        extractor.load(f)
-        extractor.project()
+        extractor.load(f).project().postProject()
         g = driver.getWholeGraph()
     }